// AMM/SushiSwapV3/types.ts
export type RemoveOptions = {
    slipage?: number
    deadline?: number
    percent?: number
    tokenId?: number
}

export type Position = {
    tokenId: number
    nonce: bigint
    operator: string
    token0: string
    token1: string
    fee: number
    tickLower: number
    tickUpper: number
    liquidity: bigint
    feeGrowthInside0LastX128: bigint
    feeGrowthInside1LastX128: bigint
    tokensOwed0: bigint
    tokensOwed1: bigint
}

export type Chains = 'arbitrum' | 'polygon' | 'optimism' | 'ethereum' | 'avalanche' | 'bsc' | 'polygonTestnet' | 'arbitrumTestnet' |
                     'avalancheTestnet' | 'ethereumTestnet'